HotFix

+++ Application/Customers/CreateCommandHandler.cs
@@ -19,15 +19,22 @@ public class CreateCommandHandler:ICustomerService
     }
     public void CreateAsync(CreateCommand command)
     {
+        CreateValidation(command);
         var customer = new Customer(command.FirstName, command.LastName, command.DateOfBirth, command.PhoneNumber,
-            command.Email, command.BankAccountNumber,_customerDomainService);
+            command.Email, command.BankAccountNumber);
          _customerRepository.CreateAsync(customer);
     
     }
 
+ 
+    
+
     public void UpdateAsync(UpdateCommand command)
     {
         var customer =  _customerRepository.GetById(command.Id);
+        if (customer == null)
+            throw new Exception("this customer is not exist");
+        UpdateValidation(command);
         _customerRepository.UpdateAsync(customer,command, _customerDomainService);
 
       
@@ -37,10 +44,27 @@ public class CreateCommandHandler:ICustomerService
     {
 
         var customer =  _customerRepository.GetById(command.Id);
-
-         customer.DeleteValidate(command.Id);
+        if (customer == null)
+            throw new Exception("this customer is not exist");
+        customer.DeleteValidate(command.Id);
          _customerRepository.DeleteAsync(customer);
-      
-       
     }
+
+    private void CreateValidation(CreateCommand command)
+    {
+        _customerDomainService.ValidateCustomer(command.FirstName, command.LastName, command.DateOfBirth, command.Email, 0);
+        _customerDomainService.ValidateEmail(command.Email);
+        _customerDomainService.ValidatePhoneNumber(command.PhoneNumber);
+        _customerDomainService.ValidateBAccNumber(command.BankAccountNumber);
+
+    }
+    private void UpdateValidation(UpdateCommand command)
+    {
+        _customerDomainService.ValidateCustomer(command.FirstName, command.LastName, command.DateOfBirth, command.Email, command.Id);
+        _customerDomainService.ValidateEmail(command.Email);
+        _customerDomainService.ValidatePhoneNumber(command.PhoneNumber);
+        _customerDomainService.ValidateBAccNumber(command.BankAccountNumber);
+
+    }
+
 }
